Fix/Refactor Profile folder structure
